Use tsc (again) to build ts files and generate .d.ts files
Use relative imports instead of absolute because tsc cannot resolve paths when generating declaration files. Since there are no absolute imports, we can use only tsc to build the source code. It's possible to use other libraries to solve the problem, but it's too much complexity for something that can be simpler for now.